requirements.txt
aiogram>=3.6,<4
aiosqlite>=0.20,<1
python-dotenv>=1.0,<2
